add expenses function and test
This function contains full crud for expenses. A complete ttd test.

+++ app/Http/Controllers/PropertyController.php
@@ -181,11 +181,13 @@ class PropertyController extends Controller
 
         $this->removeRelatedModels($property);
 
-        $property->images()->each(function ($image) {
-            Storage::delete($image->path);
+        if($property->images()) {
+            $property->images()->each(function ($image) {
+                Storage::delete($image->path);
 
-            $image->delete();
-        });
+                $image->delete();
+            });
+        }
 
         $property->delete();
     }
